requirements.txt
<<<<<<< HEAD
streamlit
pandas
openai>=1.0.0
plotly
python-dotenv